fix countdown and ux
