--- .prettierrc
@@ -1,3 +1,0 @@
-trailingComma: "es5"
-singleQuote: true
-printWidth: 120
